Add CI workflow and refactor test structure
Introduce a CI workflow. Refactor tests by moving them to the 'integration' package and setting up tag-based test inclusion/exclusion in the build process.
