+++ .github/workflows/ci.yml
@@ -1,7 +1,9 @@
-name: Check CTRF
+name: CI
 
 on:
-  workflow_dispatch:
+  pull_request:
+    branches:
+      - master
 
 jobs:
   build:
@@ -22,15 +24,11 @@ jobs:
         uses: gradle/actions/setup-gradle@v4
 
       - name: Build with Gradle Wrapper
-        run: ./gradlew clean test -PincludedTags=integration
+        run: ./gradlew clean build -x test
+
+      - name: Run unit tests
+        run: ./gradlew test -PexcludedTags=integration
 
       - name: Run CTRF annotations
         run: npx github-actions-ctrf build/test-results/test/ctrf-report.json
-        if: always()
-
-      - name: Upload test results
-        uses: actions/upload-artifact@v4
-        with:
-          name: test-results
-          path: build/test-results
         if: always()
