+++ build.gradle
@@ -30,7 +30,23 @@ dependencies {
 }
 
 test {
-    useJUnitPlatform()
+    useJUnitPlatform {
+        if (project.hasProperty("includedTags")) {
+            def includedTags = project.properties.get("includedTags").toString().split(',')
+            includedTags.each { tag ->
+                println "Including tag: $tag"
+                includeTags(tag.trim())
+            }
+        }
+
+        if (project.hasProperty("excludedTags")) {
+            def excludedTags = project.properties.get("excludedTags").toString().split(',')
+            excludedTags.each { tag ->
+                println "Excluding tag: $tag"
+                excludeTags(tag.trim())
+            }
+        }
+    }
 }
 
 checkstyle {
