+++ src/test/java/io/github/alexshamrai/BaseTest.java
@@ -1,0 +1,8 @@
+package io.github.alexshamrai;
+
+import io.github.alexshamrai.jupiter.CtrfExtension;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+@ExtendWith(CtrfExtension.class)
+public class BaseTest {
+}
