--- src/test/java/io/github/alexshamrai/functional/BaseTest.java
@@ -1,8 +1,0 @@
-package io.github.alexshamrai.functional;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import io.github.alexshamrai.jupiter.CtrfExtension;
-
-@ExtendWith(CtrfExtension.class)
-public abstract class BaseTest {
-}
