+++ src/test/java/io/github/alexshamrai/integration/DummyFailedTest.java
@@ -1,20 +1,17 @@
-package io.github.alexshamrai.functional;
+package io.github.alexshamrai.integration;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class DummyFailedTest extends BaseTest {
+public class DummyFailedTest extends BaseIntegrationTest {
 
     @Test
-    @Disabled
     void firstFailedTest() {
         System.out.println("DummyFailedTest firstFailedTest()");
         assert false;
     }
 
     @Test
-    @Disabled
     @DisplayName("Second failed test")
     void secondFailedTest() {
         System.out.println("DummyFailedTest secondFailedTest()");
